Merge remote-tracking branch 'origin/master'

Conflicts:
	qic_api/qic_include/qic_xuctrl.h
	qic_api/qic_src/qic_control.c
	qic_example/example_avconly.c
	qic_example/misc_writefile.c

+++ qic_api/qic_src/qic_api_stream.c
@@ -576,7 +576,7 @@ int qic_getframe2(unsigned int dev_id)
     struct v4l2_buffer buf;
     char *video_buffer;
     unsigned int video_buffer_len;
-    unsigned short timestamp=0;
+    unsigned long timestamp=0;
     int avc_stream_id;
     out_frame_t sframe;
     //	struct timespec ts;
